third_party/blink/web_tests/fast/forms/submit-form-with-dirname-attribute-expected.txt

Test that when dirname attribute is specified then it is added in submission body.

Comment: 

WithRTLDir: 

WithLTRDir: 

WithInvalidDir: 

WithRTLValue: 

Post Comment

PASS document.location.search.indexOf("comment.dir=ltr") != -1 is true
PASS document.location.search.indexOf("txtareaRTL.dir=rtl") != -1 is true
PASS document.location.search.indexOf("txtareaLTR.dir=ltr") != -1 is true
PASS document.location.search.indexOf("txtareaInvalid.dir=ltr") != -1 is true
PASS document.location.search.indexOf("inputRTLvalue.dir=ltr") != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
